+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4099](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4099))
 - `opentelemetry-instrumentation-aiohttp-server` Implement new semantic convention opt-in migration
   ([#3980](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3980))
+- `opentelemetry-instrumentation-sqlalchemy`, `opentelemetry-instrumentation-dbapi`: Add `commenter_for_all_spans` option to enable SQLCommenter for all spans regardless of sampling decision
+  ([#4090](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4090))
 - `opentelemetry-instrumentation-falcon`: pass request attributes at span creation
   ([#4119](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4119))
 - `opentelemetry-instrumentation`: add database stability attribute setters in `_semconv` utilities